// src/command-bundles/polls/types/errors.ts
export enum PollErrors {
  POLL_CREATION_NO_DESCRIPTION_PROVIDED= 'POLL_CREATION_NO_DESCRIPTION_PROVIDED',
  POLL_ALREADY_EXIST_IN_CHANNEL = 'POLL_ALREADY_EXIST_IN_CHANNEL',
  POLL_DOES_NOT_EXIST_IN_CHANNEL = 'POLL_DOES_NOT_EXIST_IN_CHANNEL',
  POLL_ADD_OPTION_NO_DESCRIPTION_PROVIDED = 'POLL_ADD_OPTION_NO_DESCRIPTION_PROVIDED',
  POLL_VOTE_ALREADY_EXISTS = 'POLL_VOTE_ALREADY_EXISTS',
  POLL_VOTE_DOES_NOT_EXIST = 'POLL_VOTE_DOES_NOT_EXIST',
}